created new css and components folders and moved files there for better organization, changed file paths for app.js and index.js to reflect change
